+++ tests/testthat/test-Rules-class.R
@@ -186,6 +186,22 @@ test_that("NextBestTDsamples object can be created with user constructor", {
   expect_equal(result@derive(c(1:5)), 3) # nolintr
 })
 
+## NextBestMaxGain ----
+
+test_that(".NextBestMaxGain works as expected", {
+  result <- expect_silent(.NextBestMaxGain())
+  expect_valid(result, "NextBestMaxGain")
+})
+
+test_that("NextBestMaxGain object can be created with user constructor", {
+  result <- expect_silent(
+    NextBestMaxGain(0.4, 0.35)
+  )
+  expect_valid(result, "NextBestMaxGain")
+  expect_equal(result@prob_target_drt, 0.4)
+  expect_equal(result@prob_target_eot, 0.35)
+})
+
 # Increments ----
 
 ## IncrementsRelativeDLTCurrent-class ----